test.py: refactor: split pylib/scylla_cluster.py
Move ScyllaServer and its helpers (make_scylla_conf, SCYLLA_CMDLINE_OPTIONS,
ScyllaVersionDescription and version helpers, merge_cmdline_options,
start_stop_lock, stop_event) to pylib/scylla_server.py, and
ScyllaClusterManager with the manager_op decorator to
pylib/scylla_cluster_manager.py.

Also, move async_rmtree() and io_executor to pylib/util.py and drop the
unused contextlib, fcntl and urllib imports.

Pure code motion; all imports are repointed at the new modules, in the
test files too.
